Refactor test setup and update requirements
- Removed `pytest` from `requirements.txt` as it is no longer needed.
- Simplified the `compile_datapack` fixture in `conftest.py` to improve cleanup logic and manage compiled datapack directories more effectively.
- Updated import statements in `test_compile_examples.py` to reflect changes in helper module structure.
